gradle build of instrumentation
For testing of android application, the build script of app subproject is
modified.
I could not find a way to do this by writing a gradle script inside the
instrumentation module. Instead, I had to change the app/build.gradle and
include paths for instrumentation src and manifest.

Another peculiarity was the error
"indirect dependency on Android API level 15, but minSdkVersion for variant
'Debug' is API level 7"

I haven't yet been able to resolve this 'indirect' dependency problem. One
thought is adding android appcompat library, But it conflicts with the available
actionbarsherlock version. One way of doing this would be to migrate from
actionbarsherlock to android support appcompat. But that would require some
extra efforts.

As a shortcut, I have replaced minSdkVersion in app/build.gradle to 15 and the
instrumentation works. But this is temporary respite.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>