Remove httpclient warning from android build
Because the official Android APIs includes httpclient we remove all dependency
on httpclient, including its transitive dependency. Not doing this gives the
warning :

Dependency org.apache.httpcomponents:httpclient:4.0.3 is ignored for debug as it
may be conflicting with the internal version provided by Android. In case of
problem, please repackage it with jarjar to change the class packages.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>