Throw/handle exception on Last.fm api request fail Fixes #2
The debug version uses an expired API key of lastfm. This causes the API request
to fail and give a null object.
Due to this the application was crashing in the Artist information activity.

The bug was fixed by throwing a NullPointerException if the API request fails
and null object is received. And this exception is handled in the ArtistLoader
activity.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>